requirements.txt
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.21.0,<0.22.0
pillow>=8.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
rembg>=2.0.0
numpy>=1.19.5
pinecone>=2.2.0
python-dotenv>=0.19.0
requests>=2.26.0 
onnx
onnxruntime
imagehash>=4.2.0
pillow-avif-plugin>=1.4.0
moondream
